+++ gulp/tasks/watchify.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports = function(gulp){
-  gulp.task('watchify', function(callback) {
-    // Start browserify task with devMode === true
-    //browserifyTask(callback, true);
-  });
-};