+++ test/src/index.js
@@ -1,0 +1,2 @@
+console.log('Javascript is running.')
+console.log('changr')